refactor: move shellwire under internal/shell
Relocate the shellwire package from internal/shellwire to
internal/shell/shellwire, making it a subpackage of internal/shell
following the shell consolidation in #1131.

Pure package move: the Go package name stays shellwire and only the
import path changes. No changes to the wire format, Subprotocol const,
buffer sizes, or any logic. Importers and doc-comment references updated
to the new path.
