+++ internal/shell/shellwire/shellwire.go
